Abstract over tokenizer implementations with the Tokenize trait
- WordPieceTokenizer is renamed to BertTokenizer
- The Tokenize trait is now implemented on tokenizers rather than
  Sentence.
- Use the Tokenize trait throughout the project rather a specific
  tokenizer.
