+++ sticker2-utils/src/subcommands/finetune.rs
@@ -8,7 +8,7 @@ use ordered_float::NotNan;
 use stdinout::OrExit;
 use sticker2::dataset::{ConllxDataSet, DataSet};
 use sticker2::encoders::Encoders;
-use sticker2::input::WordPieceTokenizer;
+use sticker2::input::Tokenize;
 use sticker2::lr::{ExponentialDecay, LearningRateSchedule, PlateauLearningRate};
 use sticker2::model::BertModel;
 use sticker2::optimizers::{AdamW, AdamWConfig};
@@ -101,7 +101,7 @@ impl FinetuneApp {
     fn run_epoch(
         &self,
         encoders: &Encoders,
-        tokenizer: &WordPieceTokenizer,
+        tokenizer: &dyn Tokenize,
         model: &BertModel,
         file: &mut File,
         mut optimizer: Option<&mut AdamW>,
@@ -507,7 +507,7 @@ impl StickerApp for FinetuneApp {
 
             self.run_epoch(
                 &model.encoders,
-                &model.tokenizer,
+                &*model.tokenizer,
                 &model.model,
                 &mut train_file,
                 Some(&mut opt),
@@ -520,7 +520,7 @@ impl StickerApp for FinetuneApp {
             last_acc = tch::no_grad(|| {
                 self.run_epoch(
                     &model.encoders,
-                    &model.tokenizer,
+                    &*model.tokenizer,
                     &model.model,
                     &mut validation_file,
                     None,
